Code for Analysis for Final Project
This file is a coding (Jupyter Notebook)  created in Google Colab. It provides an easy-to-follow demonstration of data analysis using Python, focusing on student performance data. The notebook is well-structured and includes clear explanations, making it simple to run and understand. Ideal for exploring student performance trends or learning basic data analysis techniques.